requirements.txt
pandas>=2.1,<3.0
numpy>=1.26,<3.0
scipy>=1.11,<2.0
plotly>=5.19,<6.0
matplotlib>=3.8,<3.10
seaborn>=0.13.0,<0.14.0
shap>=0.44,<0.47
tqdm>=4.66,<5.0
scikit-learn>=1.4,<1.6
lightgbm>=4.0,<5.0
xgboost>=2.0,<3.0
imbalanced-learn>=0.12,<0.13
joblib>=1.3,<2.0
mlflow>=2.12,<3.0
streamlit>=1.33,<2.0
python-dotenv>=1.0,<2.0

-e .